PlayPcmWin 3.0.11 seek bug fix

+++ trunk/WasapiIODLL/WasapiIOIF.cpp
@@ -403,7 +403,7 @@ WasapiIO_GetTotalFrameNum(void)
 
 extern "C" __declspec(dllexport)
 bool __stdcall
-WasapiIO_SetPosFrame(int v)
+WasapiIO_SetPosFrame(int64_t v)
 {
     assert(self);
     return self->wasapi.SetPosFrame(v);
